pyproject.toml
[tool.poetry]
name = "dit-qa-ais-pu-portals"
version = "0.1.0"
description = "Autotests for AIS-PU-PORTALS"
authors = ["Seregin Andrey <a.seregin@coms.ru>", "Sadykov Timur <t.sadykov@coms.ru>"]

packages = [
    { include = "dit" }
    ]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "4.9.1"
webdriver-manager = "^3.8.3"
coms-qa-ui = {version="^1.0.6", source='coms'}

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
allure-pytest = "^2.9.45"
pytest-sugar = "^0.9.5"
pytest-xdist = "^2.5.0"
fasteners = "^0.17.3"
black = "^22.6.0"
isort = "^5.10.1"
mypy = "^0.971"
pylint = "^2.14.5"

[[tool.poetry.source]]
name = "PyPI"
priority = "default"

[[tool.poetry.source]]
name = "coms"
url = "https://nexus.coms.ru/repository/PYPI/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
# https://docs.pytest.org/en/latest/explanation/pythonpath.html#import-modes
log_level = "DEBUG"
log_format = "%(asctime)s.%(msecs)03dZ %(levelname)-8s %(name)s %(message)s"
log_date_format = "%Y-%m-%dT%H:%M:%S"
markers = [
    "serial: run tests sequentially",
    "smoke: Smoke test-case type",
    "critical: Critical test-case priority",
    "high: High test-case priority",
    "medium: Medium test-case priority",
    "low: Low test-case priority",
]

[tool.black]
line-length = 120
skip-string-normalization = '1'
target-version = ['py310']

[tool.isort]
default_section = 'THIRDPARTY'
known_third_party = ["ais-pu-portals"]
known_application_apps = ['dit', 'tests']
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'APPLICATION_APPS', 'LOCALFOLDER']
line_length = 120
use_parentheses = true
include_trailing_comma = true
multi_line_output = 3


[tool.mypy]
python_version = "3.10"
ignore_errors = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
show_error_codes = true
disallow_untyped_calls = true
namespace_packages = true
explicit_package_bases = true

[tool.pylint.messages_control]
max-line-length = 120
disable = [
    "invalid-name",
    "missing-docstring",
    "too-many-arguments",
    "too-many-boolean-expressions",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-lines",
    "too-many-locals",
    "too-many-public-methods",
    "too-many-statements",
    "too-many-nested-blocks",
    "too-few-public-methods",
    "duplicate-code",
    "use-a-generator",
    "fixme",
    "broad-except",
]
